Add files via upload
Draw triangle, parallelogram and rectangle. 
    *                      ****               ****
   ***                    ****                ****
  *****                  ****                 ****
 *******                ****                  ****
*********              ****                   ****

Triangle (5)         Parallelogram (4x5)      Rectangle (4x5)
In the above figure between first brackets the dimension is given. For triangle the height will be given, for parallelogram and rectangle the width and height will be given. The character dot ('.') in the above figure will be represented by ASCII character space in the output. 
Input
First line of each input set indicates category C. 
If C=1 it is a triangle. Then following line of input indicates the height H ranging 1≤H≤10. 
If C=2 it is a parallelogram. Then following line of input have two integers width W and height H ranging 1≤W≤10 and 1≤H≤10. 
If C=3 it is a rectangle. Then following line of input have two integers width W and height H ranging 1≤W≤10 and 1≤H≤10. 
C= -1 indicates the end of input set. This will not be considered as input. 
Output
Output will be drawn by character '*' as following. Print a blank line after each output. 

Sample input
1
5
2
5 3
3
4 6
-1
Sample output:
    *
   ***
  *****
 *******
*********

  *****
 *****
*****

****
****
****
****
****
****